use more can/check in group.form
git-svn-id: https://forge.indepnet.net/svn/glpi/trunk@9293 35612d4c-9d10-0410-837c-bf9dd1be60d7

+++ front/group.form.php
@@ -41,47 +41,52 @@ include (GLPI_ROOT . "/inc/includes.php");
 
 if(empty($_GET["id"])) $_GET["id"] = "";
 
-$group=new Group;
+$group = new Group;
+$groupuser = new GroupUser();
+
 if (isset($_POST["add"])) {
-	$group->check(-1,'w',$_POST);
+   $group->check(-1,'w',$_POST);
+   if ($newID=$group->add($_POST)) {
+      logEvent($newID, "groups", 4, "setup", $_SESSION["glpiname"]." ".$LANG['log'][20]." ".$_POST["name"].".");
+   }
+   glpi_header($_SERVER['HTTP_REFERER']);
 
-	$newID=$group->add($_POST);
-	logEvent($newID, "groups", 4, "setup", $_SESSION["glpiname"]." ".$LANG['log'][20]." ".$_POST["name"].".");
-	glpi_header($_SERVER['HTTP_REFERER']);
 } else if (isset($_POST["delete"])) {
-	$group->check($_POST["id"],'w');
+   $group->check($_POST["id"],'w');
+   $group->delete($_POST);
+   logEvent($_POST["id"], "groups", 4, "setup", $_SESSION["glpiname"]." ".$LANG['log'][22]);
+   glpi_header($CFG_GLPI["root_doc"]."/front/group.php");
 
-	$group->delete($_POST);
-	logEvent($_POST["id"], "groups", 4, "setup", $_SESSION["glpiname"]." ".$LANG['log'][22]);
-	glpi_header($CFG_GLPI["root_doc"]."/front/group.php");
 } else if (isset($_POST["update"])) {
-	$group->check($_POST["id"],'w');
+   $group->check($_POST["id"],'w');
+   $group->update($_POST);
+   logEvent($_POST["id"], "groups", 4, "setup", $_SESSION["glpiname"]." ".$LANG['log'][21]);
+   glpi_header($_SERVER['HTTP_REFERER']);
 
-	$group->update($_POST);
-	logEvent($_POST["id"], "groups", 4, "setup", $_SESSION["glpiname"]." ".$LANG['log'][21]);
-	glpi_header($_SERVER['HTTP_REFERER']);
 } else if (isset($_POST["adduser"])) {
-	$group->check($_POST["groups_id"],'w');
+   $groupuser->check(-1,'w',$_POST);
+   if ($groupuser->add($_POST)) {
+      logEvent($_POST["groups_id"], "groups", 4, "setup", $_SESSION["glpiname"]." ".$LANG['log'][48]);
+   }
+   glpi_header($_SERVER['HTTP_REFERER']);
 
-	addUserGroup($_POST["users_id"],$_POST["groups_id"]);
-
-	logEvent($_POST["groups_id"], "groups", 4, "setup", $_SESSION["glpiname"]." ".$LANG['log'][48]);
-	glpi_header($_SERVER['HTTP_REFERER']);
 } else if (isset($_POST["deleteuser"])) {
-	$group->check($_POST["groups_id"],'w');
+   if (count($_POST["item"])) {
+      foreach ($_POST["item"] as $key => $val) {
+         if ($groupuser->can($key,'w')) {
+            deleteUserGroup($key);
+         }
+      }
+   }
+   logEvent($_POST["groups_id"], "groups", 4, "setup", $_SESSION["glpiname"]." ".$LANG['log'][49]);
+   glpi_header($_SERVER['HTTP_REFERER']);
 
-	if (count($_POST["item"]))
-		foreach ($_POST["item"] as $key => $val)
-			deleteUserGroup($key);
-
-	logEvent($_POST["groups_id"], "groups", 4, "setup", $_SESSION["glpiname"]." ".$LANG['log'][49]);
-	glpi_header($_SERVER['HTTP_REFERER']);
 } else {
-	commonHeader($LANG['Menu'][36],$_SERVER['PHP_SELF'],"admin","group");
+   commonHeader($LANG['Menu'][36],$_SERVER['PHP_SELF'],"admin","group");
 
-	$group->showForm($_SERVER['PHP_SELF'],$_GET["id"]);
+   $group->showForm($_SERVER['PHP_SELF'],$_GET["id"]);
 
-	commonFooter();
+   commonFooter();
 }
